[ftr] split cases config in 2 groups
