+++ .buildkite/ftr_configs.yml
@@ -181,7 +181,8 @@ enabled:
   - x-pack/test/functional_embedded/config.ts
   - x-pack/test/functional_enterprise_search/without_host_configured.config.ts
   - x-pack/test/functional_execution_context/config.ts
-  - x-pack/test/functional_with_es_ssl/apps/cases/config.ts
+  - x-pack/test/functional_with_es_ssl/apps/cases/group1/config.ts
+  - x-pack/test/functional_with_es_ssl/apps/cases/group2/config.ts
   - x-pack/test/functional_with_es_ssl/apps/discover_ml_uptime/config.ts
   - x-pack/test/functional_with_es_ssl/apps/triggers_actions_ui/config.ts
   - x-pack/test/functional/apps/advanced_settings/config.ts
