+++ x-pack/test/functional_with_es_ssl/apps/cases/group1/deletion.ts
@@ -5,12 +5,12 @@
  * 2.0.
  */
 
-import { FtrProviderContext } from '../../ftr_provider_context';
-import { users, roles, casesReadDeleteUser, casesAllUser, casesNoDeleteUser } from './common';
+import { FtrProviderContext } from '../../../ftr_provider_context';
+import { users, roles, casesReadDeleteUser, casesAllUser, casesNoDeleteUser } from '../common';
 import {
   createUsersAndRoles,
   deleteUsersAndRoles,
-} from '../../../cases_api_integration/common/lib/authentication';
+} from '../../../../cases_api_integration/common/lib/authentication';
 
 export default ({ getPageObjects, getService }: FtrProviderContext) => {
   const PageObjects = getPageObjects(['security', 'header']);
