+++ x-pack/test/functional_with_es_ssl/apps/cases/group2/config.ts
@@ -8,13 +8,13 @@
 import { FtrConfigProviderContext } from '@kbn/test';
 
 export default async function ({ readConfigFile }: FtrConfigProviderContext) {
-  const baseConfig = await readConfigFile(require.resolve('../../config.base.ts'));
+  const baseConfig = await readConfigFile(require.resolve('../../../config.base.ts'));
 
   return {
     ...baseConfig.getAll(),
     testFiles: [require.resolve('.')],
     junit: {
-      reportName: 'Chrome X-Pack UI Functional Tests with ES SSL - Cases',
+      reportName: 'Chrome X-Pack UI Functional Tests with ES SSL - Cases - group 2',
     },
   };
 }
