+++ x-pack/test/functional_with_es_ssl/apps/cases/group2/upgrade.ts
@@ -9,8 +9,8 @@ import expect from '@kbn/expect';
 import {
   getExternalServiceSimulatorPath,
   ExternalServiceSimulator,
-} from '../../../alerting_api_integration/common/plugins/actions_simulators/server/plugin';
-import { FtrProviderContext } from '../../ftr_provider_context';
+} from '../../../../alerting_api_integration/common/plugins/actions_simulators/server/plugin';
+import { FtrProviderContext } from '../../../ftr_provider_context';
 
 export default ({ getPageObject, getService }: FtrProviderContext) => {
   const cases = getService('cases');
